+++ src/digraph.rs
@@ -47,7 +47,7 @@ impl<N, E> DiGraph<N, E> where N: Copy + Clone + Eq + Hash
         }
     }
 
-    pub fn node_count(&self) -> uint
+    pub fn node_count(&self) -> usize
     {
         self.nodes.len()
     }
@@ -267,7 +267,7 @@ macro_rules! iterator_methods {
         }
 
         #[inline]
-        fn size_hint(&self) -> (uint, Option<uint>)
+        fn size_hint(&self) -> (usize, Option<usize>)
         {
             self.iter.size_hint()
         }
